util: add UsedAnyMove and UsedMove matchers
